% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_model}
\alias{calibrate_model}
\title{Calibrate model probabilities}
\usage{
calibrate_model(model, test_data, method = "platt")
}
\arguments{
\item{model}{caret or xgboost model}

\item{test_data}{test dataframe}

\item{method}{"platt" or "isotonic"}
}
\value{
calibrated probs
}
\description{
Calibrate model probabilities
}
