% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{get_embeddings}
\alias{get_embeddings}
\title{Get Embeddings from Autoencoder (stub)}
\usage{
get_embeddings(ae_obj, data, feature_cols = NULL)
}
\arguments{
\item{ae_obj}{Autoencoder object}

\item{data}{Input data}

\item{feature_cols}{Columns to use as features}
}
\value{
Matrix of embeddings (currently NULL since this is a stub)
}
\description{
Placeholder for extracting embeddings from a trained autoencoder.
}
