% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{prepare_model_data}
\alias{prepare_model_data}
\title{Prepare dataset for modeling}
\usage{
prepare_model_data(df, outcome_col = "Label")
}
\arguments{
\item{df}{A data.frame}

\item{outcome_col}{Name of the outcome column}
}
\value{
A processed data.frame with factor outcome
}
\description{
Prepare dataset for modeling
}
