% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_biomor.R
\name{train_biomor}
\alias{train_biomor}
\title{Train BioMoR Autoencoder}
\usage{
train_biomor(data, feature_cols, epochs = 100, batch_size = 50, lr = 0.001)
}
\arguments{
\item{data}{Dataframe with numeric features + Label}

\item{feature_cols}{Character vector of feature columns}

\item{epochs}{Number of training epochs}

\item{batch_size}{Batch size}

\item{lr}{Learning rate}
}
\value{
list(model, dataset, embeddings)
}
\description{
Train BioMoR Autoencoder
}
