% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{train_rf}
\alias{train_rf}
\title{Train a Random Forest model with caret}
\usage{
train_rf(df, outcome_col = "Label", ctrl)
}
\arguments{
\item{df}{A data.frame containing predictors and outcome}

\item{outcome_col}{Name of the outcome column (binary factor)}

\item{ctrl}{A caret::trainControl object}
}
\value{
A caret train object
}
\description{
Train a Random Forest model with caret
}
